增加了声音提示， 调整了图表线条过密的问题，之前是显示1000个点，现在100个， 数据太大的问题，可能修正了，没法复现

+++ app/src/main/java/com/chwishay/orthrecnursing/DispatchUtil.kt
@@ -46,21 +46,21 @@ object DispatchUtil {
 
     fun onResultObservable(): Observable<DataInfo> = resultSubject
 
-    var everydayTrainingDuration = 0.toUInt()
-    var everydayTrainingNum = 0.toUInt()
-    var eachGroupTrainingNum = 0.toUInt()
-    var targetJointAngle = 0.toUInt()
-    var sumTrainingDuration = 0.toShort()
-    var currentTrainingNum = 0.toShort()
+    var everydayTrainingDuration = 0
+    var everydayTrainingNum = 0
+    var eachGroupTrainingNum = 0
+    var targetJointAngle = 0
+    var sumTrainingDuration = 0
+    var currentTrainingNum = 0
     var jointAngle = 0f
     var jointAngleVelocity = 0f
-    var lateralFemoralMuscle = 0.toUInt()
-    var medialFemoris = 0.toUInt()
-    var bicepsFemoris = 0.toUInt()
-    var semitendinosusFemoris = 0.toUInt()
-    var tibialisAnteriorMuscle = 0.toUInt()
-    var peroneusLongus = 0.toUInt()
-    var exception = 0.toUInt()
+    var lateralFemoralMuscle = 0
+    var medialFemoris = 0
+    var bicepsFemoris = 0
+    var semitendinosusFemoris = 0
+    var tibialisAnteriorMuscle = 0
+    var peroneusLongus = 0
+    var exception = 0
 
 //    val timeCounterLiveData = MutableLiveData("0m0s")
 
@@ -181,37 +181,37 @@ object DispatchUtil {
     }
 
     fun resetData() {
-        everydayTrainingDuration = 0.toUInt()
-        everydayTrainingNum = 0.toUInt()
-        eachGroupTrainingNum = 0.toUInt()
-        targetJointAngle = 0.toUInt()
-        currentTrainingNum = 0.toShort()
+        everydayTrainingDuration = 0
+        everydayTrainingNum = 0
+        eachGroupTrainingNum = 0
+        targetJointAngle = 0
+        currentTrainingNum = 0
         jointAngle = 0f
         jointAngleVelocity = 0f
-        lateralFemoralMuscle = 0.toUInt()
-        medialFemoris = 0.toUInt()
-        bicepsFemoris = 0.toUInt()
-        semitendinosusFemoris = 0.toUInt()
-        tibialisAnteriorMuscle = 0.toUInt()
-        peroneusLongus = 0.toUInt()
+        lateralFemoralMuscle = 0
+        medialFemoris = 0
+        bicepsFemoris = 0
+        semitendinosusFemoris = 0
+        tibialisAnteriorMuscle = 0
+        peroneusLongus = 0
     }
 }
 
 data class DataInfo(
-    var everydayTrainingDuration: UInt = 0.toUInt(),
-    var everydayTrainingNum: UInt = 0.toUInt(),
-    var eachGroupTrainingNum: UInt = 0.toUInt(),
-    var targetJointAngle: UInt = 0.toUInt(),
-    var sumTrainingDuration: Short = 0,
-    var currentTrainingNum: Short = 0,
+    var everydayTrainingDuration: Int = 0,
+    var everydayTrainingNum: Int = 0,
+    var eachGroupTrainingNum: Int = 0,
+    var targetJointAngle: Int = 0,
+    var sumTrainingDuration: Int = 0,
+    var currentTrainingNum: Int = 0,
     var jointAngle: Float = 0f,
     var jointAngleVelocity: Float = 0f,
-    var lateralFemoralMuscle: UInt = 0.toUInt(),    //股外侧肌
-    var medialFemoris: UInt = 0.toUInt(),            //股内侧肌
-    var bicepsFemoris: UInt = 0.toUInt(),            //股二头肌
-    var semitendinosusFemoris: UInt = 0.toUInt(),    //股半腱肌
-    var tibialisAnteriorMuscle: UInt = 0.toUInt(),  //胫前肌
-    var peroneusLongus: UInt = 0.toUInt(),           //腓长肌
+    var lateralFemoralMuscle: Int = 0,    //股外侧肌
+    var medialFemoris: Int = 0,            //股内侧肌
+    var bicepsFemoris: Int = 0,            //股二头肌
+    var semitendinosusFemoris: Int = 0,    //股半腱肌
+    var tibialisAnteriorMuscle: Int = 0,  //胫前肌
+    var peroneusLongus: Int = 0,           //腓长肌
     var exceptionCode: Int = 0
 )
 
@@ -254,20 +254,20 @@ fun Float.toBytes() = java.lang.Float.floatToIntBits(this)
 
 fun ByteArray.parseData(): DataInfo {
     return DataInfo(
-        this[0].toUInt(),
-        this[1].toUInt(),
-        this[2].toUInt(),
-        this[3].toUInt(),
-        this.copyOfRange(4, 6).toShortLE(),
-        this.copyOfRange(6, 8).toShortLE(),
+        this[0].toUByte().toInt(),
+        this[1].toUByte().toInt(),
+        this[2].toUByte().toInt(),
+        this[3].toUByte().toInt(),
+        this.copyOfRange(4, 6).toShortLE().toInt(),
+        this.copyOfRange(6, 8).toShortLE().toInt(),
         this.copyOfRange(8, 10).toShortLE() / 10f,
         this.copyOfRange(10, 12).toShortLE() / 10f,
-        this[12].toUInt(),
-        this[13].toUInt(),
-        this[14].toUInt(),
-        this[15].toUInt(),
-        this[16].toUInt(),
-        this[17].toUInt(),
+        this[12].toUByte().toInt(),
+        this[13].toUByte().toInt(),
+        this[14].toUByte().toInt(),
+        this[15].toUByte().toInt(),
+        this[16].toUByte().toInt(),
+        this[17].toUByte().toInt(),
         this[18].toInt()
     )
 }
